requirements.txt
gradio==3.50.2
transformers
diffusers
torch
Pillow
